Rename: 전시회목록 api 파일명 다른 api파일들과 통일감있게 수정
